requirements/test.txt
coverage==7.6.1
factory_boy==3.3.0
Faker==35.0.0
pytest==8.3.2
pytest-cov==6.0.0
pytest-mock==3.14.0
pytest-xdist==3.6.1
python3-cyberfusion-common==2.11
requests-mock==1.12.1
